refactor: move class to action domain
